Make filename Resolution value an int. Arises due to python 3 float division Amend in old code also.

+++ eniric/original_code/nIRanalysis.py
@@ -147,7 +147,7 @@ def convolution(spectrum, band, vsini, R, epsilon=0.6, FWHM_lim=5.0, plot=True):
     #Note: difference in sampling at 1.0 and 1.5 microns makes jumps in the beginning of Y and H bands
 
     name_model = name_assignment(spectrum)
-    filename = results_dir+"Spectrum_"+name_model+"_"+band+"band_vsini"+str(vsini)+"_R"+str(R/1000)+"k.txt"
+    filename = results_dir+"Spectrum_"+name_model+"_"+band+"band_vsini"+str(vsini)+"_R"+str(int(R/1000))+"k.txt"
     write_3col(filename, wav_band, flux_band, flux_conv_res)
     print("Done.")
